kparts: query the part to close from KParts::MainWindow
because KParts::ReadWritePart::closeUrl() is not called from the
KParts::ReadWritePart destructor (and even if it is bad stuff may happen)
for most rw parts (exception being okular which does its own query, maybe
some other parts too) saving was not actually done (e.g. for ark). now the
part will be queried to close, if the upload job fails (e.g. because of
insufficient access) the application/part will not close (intentionally)
tho and no error is reported via message box which is something that needs
to be looked into

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>